pyproject.toml
[project]
name = "hiddenlayer-openai-guardrails"
version = "0.2.0"
description = "Guardrails for the OpenAI Agents SDK"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    { name = "HiddenLayer", email = "sdks@hiddenlayer.com" },
]
dependencies = [
    "hiddenlayer-sdk>=3.3.0",
    "openai-agents>=0.6.1",
]
requires-python = ">= 3.10"
classifiers = [
  "Typing :: Typed",
  "Intended Audience :: Developers",
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Operating System :: OS Independent",
  "Operating System :: POSIX",
  "Operating System :: MacOS",
  "Operating System :: POSIX :: Linux",
  "Operating System :: Microsoft :: Windows",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: Apache Software License"
]

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "fastmcp>=2.14.5",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"
extend-exclude = [
    ".venv",
    "hiddenlayer-sdk-python-3.3.0",
    "test_rig",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["hiddenlayer_openai_guardrails"]
combine-as-imports = true

[tool.pytest.ini_options]
norecursedirs = [
    ".venv",
    "hiddenlayer-sdk-python-3.3.0",
    "test_rig",
]
markers = [
    "unit: deterministic unit tests with mocked external dependencies",
    "integration: live integration tests that call external services",
]
